chore(django): refactor django response wrappers
