+++ ddtrace/contrib/internal/django/_asgi.py
@@ -1,7 +1,13 @@
 """
 Module providing async hooks. Do not import this module unless using Python >= 3.6.
 """
-from ddtrace.contrib import trace_utils
+
+from types import FunctionType
+from typing import Any
+from typing import Dict
+from typing import Tuple
+
+from ddtrace._trace.pin import Pin
 from ddtrace.contrib.internal.asgi.middleware import span_from_scope
 from ddtrace.contrib.internal.django.utils import REQUEST_DEFAULT_RESOURCE
 from ddtrace.contrib.internal.django.utils import _after_request_tags
@@ -9,8 +15,7 @@ from ddtrace.contrib.internal.django.utils import _before_request_tags
 from ddtrace.internal.utils import get_argument_value
 
 
-@trace_utils.with_traced_module
-async def traced_get_response_async(django, pin, func, instance, args, kwargs):
+async def traced_get_response_async(func: FunctionType, args: Tuple[Any, ...], kwargs: Dict[str, Any]) -> Any:
     """Trace django.core.handlers.base.BaseHandler.get_response() (or other implementations).
 
     This is the main entry point for requests.
@@ -18,7 +23,11 @@ async def traced_get_response_async(django, pin, func, instance, args, kwargs):
     Django requests are handled by a Handler.get_response method (inherited from base.BaseHandler).
     This method invokes the middleware chain and returns the response generated by the chain.
     """
-    request = get_argument_value(args, kwargs, 0, "request")
+    instance = args[0]
+
+    pin = Pin.get_from(instance)
+
+    request = get_argument_value(args, kwargs, 1, "request")
     span = span_from_scope(request.scope)
     if span is None:
         return await func(*args, **kwargs)
