+++ ddtrace/contrib/internal/django/patch.py
@@ -8,7 +8,6 @@ specific Django apps like Django Rest Framework (DRF).
 """
 
 from collections.abc import Iterable
-import functools
 from inspect import getmro
 from inspect import unwrap
 import os
@@ -19,15 +18,10 @@ import wrapt
 from wrapt.importer import when_imported
 
 from ddtrace import config
-from ddtrace.constants import SPAN_KIND
 from ddtrace.contrib import trace_utils
 from ddtrace.contrib.internal.django.user import _DjangoUserInfoRetriever
-from ddtrace.contrib.internal.trace_utils import _get_request_header_user_agent
-from ddtrace.ext import SpanKind
 from ddtrace.ext import SpanTypes
-from ddtrace.ext import http
 from ddtrace.internal import core
-from ddtrace.internal._exceptions import BlockingException
 from ddtrace.internal.constants import COMPONENT
 from ddtrace.internal.core.event_hub import ResultType
 from ddtrace.internal.logger import get_logger
@@ -36,9 +30,6 @@ from ddtrace.internal.schema import schematize_url_operation
 from ddtrace.internal.schema.span_attribute_schema import SpanDirection
 from ddtrace.internal.telemetry import get_config as _get_config
 from ddtrace.internal.utils import get_argument_value
-from ddtrace.internal.utils import get_blocked
-from ddtrace.internal.utils import http as http_utils
-from ddtrace.internal.utils import set_blocked
 from ddtrace.internal.utils.formats import asbool
 from ddtrace.internal.utils.importlib import func_name
 from ddtrace.settings.asm import config as asm_config
@@ -271,146 +262,6 @@ def traced_load_middleware(django, pin, func, instance, args, kwargs):
     return func(*args, **kwargs)
 
 
-def _gather_block_metadata(request, request_headers, ctx: core.ExecutionContext):
-    from . import utils
-
-    try:
-        metadata = {http.STATUS_CODE: "403", http.METHOD: request.method}
-        url = utils.get_request_uri(request)
-        query = request.META.get("QUERY_STRING", "")
-        if query and config_django.trace_query_string:
-            metadata[http.QUERY_STRING] = query
-        user_agent = _get_request_header_user_agent(request_headers)
-        if user_agent:
-            metadata[http.USER_AGENT] = user_agent
-    except Exception as e:
-        log.warning("Could not gather some metadata on blocked request: %s", str(e))
-    core.dispatch("django.block_request_callback", (ctx, metadata, config_django, url, query))
-
-
-def _block_request_callable(request, request_headers, ctx: core.ExecutionContext):
-    # This is used by user-id blocking to block responses. It could be called
-    # at any point so it's a callable stored in the ASM context.
-    from django.core.exceptions import PermissionDenied
-
-    set_blocked()
-    _gather_block_metadata(request, request_headers, ctx)
-    raise PermissionDenied()
-
-
-@trace_utils.with_traced_module
-def traced_get_response(django, pin, func, instance, args, kwargs):
-    """Trace django.core.handlers.base.BaseHandler.get_response() (or other implementations).
-
-    This is the main entry point for requests.
-
-    Django requests are handled by a Handler.get_response method (inherited from base.BaseHandler).
-    This method invokes the middleware chain and returns the response generated by the chain.
-    """
-    from ddtrace.contrib.internal.django.compat import get_resolver
-
-    from . import utils
-
-    request = get_argument_value(args, kwargs, 0, "request")
-    if request is None:
-        return func(*args, **kwargs)
-
-    request_headers = utils._get_request_headers(request)
-
-    with core.context_with_data(
-        "django.traced_get_response",
-        remote_addr=request.META.get("REMOTE_ADDR"),
-        headers=request_headers,
-        headers_case_sensitive=True,
-        span_name=schematize_url_operation("django.request", protocol="http", direction=SpanDirection.INBOUND),
-        resource=utils.REQUEST_DEFAULT_RESOURCE,
-        service=trace_utils.int_service(pin, config_django),
-        span_type=SpanTypes.WEB,
-        tags={COMPONENT: config_django.integration_name, SPAN_KIND: SpanKind.SERVER},
-        integration_config=config_django,
-        distributed_headers=request_headers,
-        activate_distributed_headers=True,
-        pin=pin,
-    ) as ctx, ctx.span:
-        core.dispatch(
-            "django.traced_get_response.pre",
-            (
-                functools.partial(_block_request_callable, request, request_headers, ctx),
-                ctx,
-                request,
-                utils._before_request_tags,
-            ),
-        )
-
-        response = None
-
-        def blocked_response():
-            from django.http import HttpResponse
-
-            block_config = get_blocked() or {}
-            desired_type = block_config.get("type", "auto")
-            status = block_config.get("status_code", 403)
-            if desired_type == "none":
-                response = HttpResponse("", status=status)
-                location = block_config.get("location", "")
-                if location:
-                    response["location"] = location
-            else:
-                ctype = block_config.get("content-type", "application/json")
-                content = http_utils._get_blocked_template(ctype)
-                response = HttpResponse(content, content_type=ctype, status=status)
-                response.content = content
-                response["Content-Length"] = len(content.encode())
-            utils._after_request_tags(pin, ctx.span, request, response)
-            return response
-
-        try:
-            if get_blocked():
-                response = blocked_response()
-                return response
-
-            query = request.META.get("QUERY_STRING", "")
-            uri = utils.get_request_uri(request)
-            if uri is not None and query:
-                uri += "?" + query
-            resolver = get_resolver(getattr(request, "urlconf", None))
-            if resolver:
-                try:
-                    path = resolver.resolve(request.path_info).kwargs
-                    log.debug("resolver.pattern %s", path)
-                except Exception:
-                    path = None
-
-            core.dispatch(
-                "django.start_response", (ctx, request, utils._extract_body, utils._remake_body, query, uri, path)
-            )
-            core.dispatch("django.start_response.post", ("Django",))
-
-            if get_blocked():
-                response = blocked_response()
-                return response
-
-            try:
-                response = func(*args, **kwargs)
-            except BlockingException as e:
-                set_blocked(e.args[0])
-                response = blocked_response()
-                return response
-
-            if get_blocked():
-                response = blocked_response()
-                return response
-
-            return response
-        finally:
-            core.dispatch("django.finalize_response.pre", (ctx, utils._after_request_tags, request, response))
-            if not get_blocked():
-                core.dispatch("django.finalize_response", ("Django",))
-                if get_blocked():
-                    response = blocked_response()
-                    return response  # noqa: B012
-
-
 def instrument_view(django, view, path=None):
     """
     Helper to wrap Django views.
@@ -638,14 +489,9 @@ def _patch(django):
 
     @when_imported("django.core.handlers.base")
     def _(m):
-        import django
+        from .response import instrument_module
 
-        trace_utils.wrap(m, "BaseHandler.get_response", traced_get_response(django))
-        if django.VERSION >= (3, 1):
-            # Have to inline this import as the module contains syntax incompatible with Python 3.5 and below
-            from ._asgi import traced_get_response_async
-
-            trace_utils.wrap(m, "BaseHandler.get_response_async", traced_get_response_async(django))
+        instrument_module(django)
 
     @when_imported("django.contrib.auth")
     def _(m):
@@ -720,8 +566,6 @@ def patch():
 def _unpatch(django):
     trace_utils.unwrap(django.apps.registry.Apps, "populate")
     trace_utils.unwrap(django.core.handlers.base.BaseHandler, "load_middleware")
-    trace_utils.unwrap(django.core.handlers.base.BaseHandler, "get_response")
-    trace_utils.unwrap(django.core.handlers.base.BaseHandler, "get_response_async")
     trace_utils.unwrap(django.template.base.Template, "render")
     trace_utils.unwrap(django.conf.urls.static, "static")
     trace_utils.unwrap(django.conf.urls, "url")
@@ -740,6 +584,10 @@ def _unpatch(django):
         from . import templates
 
         templates.uninstrument_module(django.template.base)
+
+    from .response import uninstrument_module
+
+    uninstrument_module(django)
 
 
 def unpatch():
